Add support for opting-out from shadowRoot behavior.
* Adjust the Readme.md.
* Bump the version.
* Add .editorsconfig file.
* Added type definition and dev dependencies
* Transpile from ES6 to ES5
