+++ index.d.ts
@@ -1,0 +1,5 @@
+declare namespace ReactWebComponent {
+  export const create: (app: JSX.Element, tagName: string, optOutFromShadowRoot?: boolean) => void;
+}
+
+export default ReactWebComponent;
