+++ src/getStyleElementsFromReactWebComponentStyleLoader.js
@@ -1,7 +1,1 @@
-module.exports = function getStyleElementsFromReactWebComponentStyleLoader() {
-  try {
-    return require('react-web-component-style-loader/exports').styleElements;
-  } catch (e) {
-    return [];
-  }
-};
+'use strict';module.exports=function(){try{return require('react-web-component-style-loader/exports').styleElements}catch(a){return[]}};
